# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mechanical Engineer Portfolio API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Project {
  id: number;
  title: string;
  description: string;
  /** @nullable */
  longDescription?: string | null;
  category: string;
  tags?: string[];
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  githubUrl?: string | null;
  /** @nullable */
  liveUrl?: string | null;
  featured: boolean;
  /** @nullable */
  order?: number | null;
  createdAt: string;
}

export interface ProjectInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  description: string;
  longDescription?: string;
  /** @minLength 1 */
  category: string;
  tags?: string[];
  imageUrl?: string;
  githubUrl?: string;
  liveUrl?: string;
  featured?: boolean;
  order?: number;
}

export interface ProjectUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  longDescription?: string;
  category?: string;
  tags?: string[];
  imageUrl?: string;
  githubUrl?: string;
  liveUrl?: string;
  featured?: boolean;
  order?: number;
}

export type ProjectStatsCategoriesItem = {
  name: string;
  count: number;
};

export interface ProjectStats {
  totalProjects: number;
  featuredCount: number;
  categories: ProjectStatsCategoriesItem[];
}

export interface AdminCredentials {
  password: string;
}

export interface AdminSession {
  isAdmin: boolean;
}

export interface ResumeInfo {
  hasResume: boolean;
  /** @nullable */
  downloadUrl: string | null;
  /** @nullable */
  fileName?: string | null;
  /** @nullable */
  updatedAt?: string | null;
}