# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Mechanical Engineer Portfolio API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: projects
    description: Portfolio project operations
  - name: admin
    description: Admin authentication operations
  - name: resume
    description: Resume operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "401":
          description: Unauthorized

  /projects/featured:
    get:
      operationId: listFeaturedProjects
      tags: [projects]
      summary: List featured/pinned projects
      responses:
        "200":
          description: Featured projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"

  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a single project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    put:
      operationId: updateProject
      tags: [projects]
      summary: Update a project (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "401":
          description: Unauthorized
        "404":
          description: Not found
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "404":
          description: Not found

  /projects/stats:
    get:
      operationId: getProjectStats
      tags: [projects]
      summary: Get portfolio stats (total projects, categories, etc.)
      responses:
        "200":
          description: Portfolio statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectStats"

  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCredentials"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Invalid credentials

  /admin/logout:
    post:
      operationId: adminLogout
      tags: [admin]
      summary: Admin logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"

  /admin/me:
    get:
      operationId: getAdminMe
      tags: [admin]
      summary: Check admin session status
      responses:
        "200":
          description: Session info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"

  /resume:
    get:
      operationId: getResume
      tags: [resume]
      summary: Get resume metadata
      responses:
        "200":
          description: Resume metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumeInfo"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Project:
      type: object
      required: [id, title, description, category, createdAt, featured]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        longDescription:
          type: ["string", "null"]
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        imageUrl:
          type: ["string", "null"]
        githubUrl:
          type: ["string", "null"]
        liveUrl:
          type: ["string", "null"]
        featured:
          type: boolean
        order:
          type: ["integer", "null"]
        createdAt:
          type: string

    ProjectInput:
      type: object
      required: [title, description, category]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        longDescription:
          type: string
        category:
          type: string
          minLength: 1
        tags:
          type: array
          items:
            type: string
        imageUrl:
          type: string
        githubUrl:
          type: string
        liveUrl:
          type: string
        featured:
          type: boolean
        order:
          type: integer

    ProjectUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        longDescription:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        imageUrl:
          type: string
        githubUrl:
          type: string
        liveUrl:
          type: string
        featured:
          type: boolean
        order:
          type: integer

    ProjectStats:
      type: object
      required: [totalProjects, categories, featuredCount]
      properties:
        totalProjects:
          type: integer
        featuredCount:
          type: integer
        categories:
          type: array
          items:
            type: object
            required: [name, count]
            properties:
              name:
                type: string
              count:
                type: integer

    AdminCredentials:
      type: object
      required: [password]
      properties:
        password:
          type: string

    AdminSession:
      type: object
      required: [isAdmin]
      properties:
        isAdmin:
          type: boolean

    ResumeInfo:
      type: object
      required: [hasResume, downloadUrl]
      properties:
        hasResume:
          type: boolean
        downloadUrl:
          type: ["string", "null"]
        fileName:
          type: ["string", "null"]
        updatedAt:
          type: ["string", "null"]