# lib/api-zod/src/generated/types/project.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mechanical Engineer Portfolio API
 * OpenAPI spec version: 0.1.0
 */

export interface Project {
  id: number;
  title: string;
  description: string;
  /** @nullable */
  longDescription?: string | null;
  category: string;
  tags?: string[];
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  githubUrl?: string | null;
  /** @nullable */
  liveUrl?: string | null;
  featured: boolean;
  /** @nullable */
  order?: number | null;
  createdAt: string;
}