# lib/api-zod/src/generated/types/projectStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mechanical Engineer Portfolio API
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectStatsCategoriesItem } from './projectStatsCategoriesItem';

export interface ProjectStats {
  totalProjects: number;
  featuredCount: number;
  categories: ProjectStatsCategoriesItem[];
}