# lib/api-zod/src/generated/types/projectUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mechanical Engineer Portfolio API
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  longDescription?: string;
  category?: string;
  tags?: string[];
  imageUrl?: string;
  githubUrl?: string;
  liveUrl?: string;
  featured?: boolean;
  order?: number;
}